pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "UnitCraft"
version = "0.1.0"
description = "This will be a Python package for UnitCraft, a tool for creating unittest cases for Python code using LLMs."
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "absl-py>=2.3.1",
    "coverage>=7.9.2",
    "google-generativeai>=0.8.5",
    "langchain>=0.3.26",
    "langchain-community>=0.3.27",
    "langchain-core>=0.3.69",
    "langchain-google-genai>=2.0.10",
    "pytest>=8.4.1",
    "pytest-cov>=6.2.1",
    "pytest-mock>=3.14.1",
    "pytest-xdist>=3.8.0",
    "python-dotenv>=1.1.1",
    "pyyaml>=6.0.2",
    "tensorflow>=2.20.0",
]

[project.scripts]
unitcraft = "main:main"

[project.optional-dependencies]
openai = ["langchain-openai>=0.1.0"]
anthropic = ["langchain-anthropic>=0.1.0"]
community = ["langchain-community>=0.3.0"]
all = [
    "langchain-openai>=0.1.0",
    "langchain-anthropic>=0.1.0",
    "langchain-community>=0.3.0"
]

[tool.setuptools]
packages = ["agents", "helpers"]
py-modules = ["main", "config"]
